+++ src/main/java/com/jss/osiris/modules/osiris/quotation/service/CentralPayDelegateService.java
@@ -8,7 +8,7 @@ public interface CentralPayDelegateService {
 
         public CentralPayPaymentRequest cancelPaymentRequest(String centralPayPaymentRequestId);
 
-        public CentralPayPaymentRequest generatePayPaymentRequest(Float amount, String mail, String entityId,
+        public CentralPayPaymentRequest generatePayPaymentRequest(Double amount, String mail, String entityId,
                         String subject);
 
         public CentralPayPaymentRequest getPaymentRequest(String centralPayPaymentRequestId) throws OsirisException;
