+++ src/main/java/com/jss/osiris/modules/osiris/tiers/model/IRffCompute.java
@@ -9,9 +9,9 @@ public interface IRffCompute {
 
     Integer getRffId();
 
-    Float getRffAl();
+    Double getRffAl();
 
-    Float getRffFor();
+    Double getRffFor();
 
     LocalDate getStartDate();
